+++ src/main/kotlin/com/github/lynxie/oxocarbon/activities/OxocarbonStartupActivity.kt
@@ -27,7 +27,4 @@ class OxocarbonStartupActivity : ProjectActivity {
             SettingsNotifications.notifyWelcome(project)
         }
     }
-
-    
-    
 }