+++ src/main/kotlin/com/github/lynxie/oxocarbon/settings/theme/ThemeVariant.kt
@@ -1,6 +1,0 @@
-package com.github.lynxie.oxocarbon.settings.theme
-
-enum class ThemeVariant(val themeName: String) {
-    DARK("Oxocarbon Dark"),
-    LIGHT("Oxocarbon Light"),
-}