+++ src/main/kotlin/com/github/lynxie/oxocarbon/topics/DropdownItemChangeListener.kt
@@ -1,0 +1,12 @@
+package com.github.lynxie.oxocarbon.topics
+
+import com.intellij.util.messages.Topic
+
+fun interface DropdownItemChangeListener {
+
+    companion object {
+        val DROPDOWN_ITEM_TOPIC : Topic<DropdownItemChangeListener> = Topic.create("Dropdown Item ChangeListener", DropdownItemChangeListener::class.java)
+    }
+
+    fun dropdownItemChanged()
+}